# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Claimbase Expense Claim Management System API
servers:
  - url: /api
    description: Base API path
security:
  - TelegramAuth: []
...
paths:
...
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [users]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
  /auth/register:
    post:
      operationId: authRegister
      tags: [users]
      summary: Register user from Telegram initData
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
  /auth/telegram-login:
    post:
      operationId: authTelegramLogin
      tags: [users]
      summary: Login via Telegram Login Widget
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramLoginResponse"
        "401":
          description: Invalid login data
  /users/telegram/{telegramId}:
    get:
      operationId: getUserByTelegramId
      tags: [users]
      summary: Get user by Telegram ID
      security:
        - InternalAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      security:
        - InternalAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
components:
  schemas:
...
    CreateUserBody:
      type: object
      properties:
        telegramId:
          type: string
        name:
          type: string
        timezone:
          type: string
      required:
        - telegramId
        - name
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        dashboardCurrency:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        name:
          type: string
        timezone:
          type: string
        dashboardCurrency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - telegramId
        - name
        - timezone
        - dashboardCurrency
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath
    TelegramLoginRequest:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        photo_url:
          type: string
        auth_date:
          type: string
        hash:
          type: string
      required:
        - id
        - first_name
        - auth_date
        - hash
    TelegramLoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - token
        - user
  securitySchemes:
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram Web App initData string (HMAC-verified, 1-hour TTL)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT session token from Telegram Login Widget auth
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Server-to-server authentication key (used by Telegram bot only)